[+] add player & move player & animate

+++ include/rpg.h
@@ -66,6 +66,6 @@ void play_sound(scene *d, enum sound_grp grp);
 void free_game(screen *hub);
 void events_manage(scene *d, sfEvent);
 void data_play(screen *screen1);
-sfVector2f create_vector(int x, int y);
+sfVector2f create_vector(float x, float y);
 
 #endif
