+++ utils/utils.c
@@ -17,7 +17,7 @@ int get_next_index(node *list)
     return i;
 }
 
-sfVector2f create_vector(int x, int y)
+sfVector2f create_vector(float x, float y)
 {
     sfVector2f v = {x, y};
     return v;
